feat: [#60] typescript 적용
