+++ FE/src/components/Login/AuthRoute.tsx
@@ -1,19 +1,23 @@
 import { useEffect } from "react";
 import { Route, Redirect } from "react-router-dom";
 import { useRecoilState } from "recoil";
-// import { LoginState } from "../../atoms/atoms";
-import { useRecoilValue } from "recoil";
 import { LoginState } from "../../atoms/atoms";
-const AuthRoute = ({ component: Component, render, ...rest }) => {
+
+// type Props = RouteProps & {
+//   component: React.ComponentType<RouteComponentProps>;
+//   render: (props: any) => React.ComponentType<RouteComponentProps>;
+//   path: string;
+// };
+const AuthRoute = ({ component: Component, render, ...rest }: any) => {
   const [, setIsLogin] = useRecoilState(LoginState);
   const token = localStorage.getItem("token");
   useEffect(() => {
     if (token !== null) setIsLogin((v) => true);
-  }, [token]);
+  }, [token, setIsLogin]);
   return (
     <Route
       {...rest}
-      render={(props) =>
+      render={(props: any) =>
         token !== null ? (
           render ? (
             render(props)
